RDK-35868 [DAC] Add unit tests to BundleGen (#70)
* RDK-35868 [DAC] Add unit tests to BundleGen

1. Added unit_tests folder which consists of L1 testing.
2. In L1 testing, It will test the api's present in bundle_processor.py script.

* Update bundle_processor.py

* Update bundle_processor.py

* Update bundle_processor.py

* Update bundle_processor.py

* Update run_L1_test.sh

* RDK-35868 [DAC] Add unit tests to BundleGen

* Update getting_test_results.py

* RDK-35868 [DAC] Add unit tests to BundleGen

1. Added unit_tests folder which consists of L1 testing.
2. In L1 testing, It will test the api's present in bundle_processor.py script.

* Update bundle_processor.py

* Update bundle_processor.py

* Update bundle_processor.py

* Update bundle_processor.py

* Update run_L1_test.sh

* RDK-35868 [DAC] Add unit tests to BundleGen

* Update getting_test_results.py

* RDK-35868 [DAC] Add unit tests to BundleGen

* RDK-35868 [DAC] Add unit tests to BundleGen

+++ bundlegen/cli/main.py
@@ -152,6 +152,9 @@ def generate(image, outputdir, platform, searchpath, creds, ipk, appmetadata, ye
     # Begin processing. Work in the output dir where the img was unpacked to
     processor = BundleProcessor(
         selected_platform.get_config(), outputdir, app_metadata_dict, nodepwalking, libmatchingmode, createmountpoints)
+    if processor == False:
+        sys.exit(1)
+
     if not processor.check_compatibility():
         # Not compatible - delete any work done so far
         shutil.rmtree(outputdir)
